Add X-Developer-Key header
Let's include public key information in the header as well, for
informational purposes. Obviously, we won't use it for validation, but
since most of these messages will be logged in public-inbox, it provides
a nice trail of historical key usage data.

Additionally, we may want to implement some kind of trust-on-first-use
setup in the future, and this allows us easy access to key information
in every message.

The header is intentionally unsigned.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>