Add most of the contributor docs
Will need to duplicate most of it to b4, but it's good to have it in
both locations. B4 could be an overkill for someone just looking to
validate a couple of patches.

I may change my mind later and just move all of the maintainer docs into
b4, once they take final shape. :)

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>