Sign Message-Id header if present
It is useful to sign the message-id header, because it is frequently
used as the patch identifier. Unfortunately, unless git-format-patch is
run with --thread, the message-id won't be generated until *after* the
sendemail-validate hook is invoked, so most of the time we won't end up
signing that header.

However, having this as an option is handy.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>