Always set the i= field
If the From: header matched the identity on the key, we skipped adding
the i= field to the signature, however there are good reasons to always
have it in place. For example, a remailer service could be replacing
the From header with their own list header and moving the original From:
into X-Original-From. If we don't have the i= field present, then we
have no easy way to figure out who the signature belongs to.

Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>