hook: refuse to run on cover letter templates
When git-send-email is invoked with --compose, it will create a cover
letter template that is not a valid RFC2822 message by itself (e.g. it
has lines starting with "GIT: " that will be removed before the message
is sent). Refuse to sign such templated messages, because the contents
are going to be modified before the message is sent and the signature is
going to be invalid anyway.

Reported-by: Paul Moore <paul@paul-moore.com>
Link: https://github.com/stacked-git/stgit/issues/414
Signed-off-by: Konstantin Ryabitsev <konstantin@linuxfoundation.org>